requirements.txt
Flask
	mysql-connector-python
	flask_mysqldb
	mysqlclient
	pyyaml==3.13b1
